+++ c/src/hashtree.c
@@ -45,7 +45,7 @@
 #endif /* UNIT_TESTING */
 
 /* Sort based on the keys in a key data tuple. */
-static int16_t key_data_compare(void *key_data_a, void *key_data_b)
+static int32_t key_data_compare(void *key_data_a, void *key_data_b)
 {
     KeyData *kd_a;
     KeyData *kd_b;
@@ -134,10 +134,10 @@ static TreeNode* create_tree_leaf_node(Hashtree *ht, TreeNode
 Hashtree* tree_create(uint16_t threshold,
         uint16_t(*hash_key)(void *),
         uint16_t(*hash_key_list)(void *),
-        int16_t(*key_compare)(void *,void *),
+        int32_t(*key_compare)(void *,void *),
         void*(*key_copy)(void *),
         void(*key_free)(void *),
-        int16_t(*data_compare)(void *,void *),
+        int32_t(*data_compare)(void *,void *),
         void*(*data_copy)(void *),
         void(*data_free)(void *))
 {
