+++ c/src/uint16_list.h
@@ -37,7 +37,7 @@
 
 
 /* Compare two uint16_t values. */
-int16_t uint16_compare(void *int_a, void *int_b);
+int32_t uint16_compare(void *int_a, void *int_b);
 
 
 /* Create a copy of a uint16_t value. */
@@ -54,7 +54,7 @@ uint16_t uint16_hash(void *key);
 
 /* Compare two lists of uint16_t values. This can be used to sort lists
  * of uint16_t lists.*/
-int16_t uint16_list_compare(void *data_a, void* data_b);
+int32_t uint16_list_compare(void *data_a, void* data_b);
 
 
 /* Copy a list of uint16_t values. */
